improve errors

+++ src/lib.rs
@@ -31,13 +31,13 @@ pub enum TileCell {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct TileInfo {
     pub name: String, //nm
-    pub size: Vec<usize>, //sz
+    pub size: Vec<i32>, //sz
     pub specs: Vec<TileCell>, //specs
     pub specs2: Option<Vec<TileCell>>, //specs2
     pub tile_type: TileType, //tp
-    pub repeat_layers: Option<Vec<usize>>, //repeatL
-    pub buffer_tiles: usize, //bfTiles
-    pub random_vars: Option<usize>, //rnd
-    pub preview_pos: usize, //ptPos
+    pub repeat_layers: Option<Vec<i32>>, //repeatL
+    pub buffer_tiles: i32, //bfTiles
+    pub random_vars: Option<i32>, //rnd
+    pub preview_pos: i32, //ptPos
     pub tags: Vec<String>, //tags
 }
